Step Four: Create a scheduled call to a rake task.
The rake task is where the real work is done. You want the code in
libexec/customer_onboarding-daemon.rb to be as light as possible, since
it's what will be growing (or not) the memory footprint of your daemon.

Calling rake is perfect because it will run in a separate process and
that process will die once the work is complete. So you don't have to
worry about that memory.